Move pod state count report to different loop
Because the routine to report pod state count is in the same loop as
the watcher's main operations (EnsureCorrectRooms/RemoveDeadRooms/AutoScale) it
can not be reported for some time while these operations are running and these are
the times the metric is needed most.

The routine was moved to another goroutine running alongside the report of GRU's
room state.