[HealthController] Perform rolling update of scheduler versions (#618)
* refactor(switchversion): do not replace pods

The switch version operation will just update the scheduler version.
Let healthcontroller perform the rolling update

* refactor(healthcontroller): perform rolling update

It is now responsibility of the healthcontroller to perform a rolling
update. When there was a switch on the current scheduler active version
we need to rollout pods with the new version. Health controller either
does autoscaling or rolling update.

On rolling update, the following will happen:
1. Are there rooms with a previous scheduler version? If so, start
   update
2. Compute how many rooms we can surge
3. Enqueue priority operation to create this rooms
4. Check how many Ready rooms we have above desired from autoscaling
5. Enqueue operation to delete those rooms - they are a buffer, should
   never offend readyTarget

* docs: rolling update scenarios

Add reference to the new Rolling Update mechanism performed by the
health_controller operation