refactor(healthcontroller): split logic of getting desired rooms
The logic to get the desired number of rooms should not depend
on the logic of ensuring this desired amount. Ideally, the desired
logic would stay in a separate worker and a separate operation
that we are going to call SchedulerController. Then, HealthController
will only be responsible for enqueueing the necessary operations
to get the current state equal to the desired. Thus, rolling
updates should NOT add any operation to the queue