refactor(healthcontroller): generate only one add/remove operation per cycle (#636)
* refactor(healthcontroller): split logic of getting desired rooms

The logic to get the desired number of rooms should not depend
on the logic of ensuring this desired amount. Ideally, the desired
logic would stay in a separate worker and a separate operation
that we are going to call SchedulerController. Then, HealthController
will only be responsible for enqueueing the necessary operations
to get the current state equal to the desired. Thus, rolling
updates should NOT add any operation to the queue

* feat(removeRooms): sort by scheduler active version

After sorting by status, sort based on scheduler's active version moving
the current available rooms (Pending, Ready and Occupied) to the bottom
of the list. Now, we'll delete first Pending, Ready and Occupied from a
non-active scheduler version to ensure version updates.

* test(healthcontroller): add complete rolling update cycles

Create a matrix of test cases with table tests, validated in spreadsheets
and with the desired behavior that each health_controller operation should
take during a rolling update. There are 4 base cases: stable occupancy,
occupancy increased, occupancy decreased, stable occupancy but only 75% of
new rooms becoming active.

* refactor(healthcontroller): only append to history if tookAction

If the operation did not took action, do not append to history. This will avoid bloat
of operation history when listing, simplifying debug.

* fix(runtime): let clientSet infere object namespace

When sending Kubernetes events, let the kubernetes clientSet infere
what is the namespace of the object (Pod) instead of having one
recorder per namespace.

* fix(runtime): invert reason and msg on k8s events