refactor(disruption): use fixed maxUnavailable (#639)
Some schedulers might decide on a high readyTarget, which means a low amount of
occupied rooms in regards to total rooms, this causes the PDB to have a
relative low number in compared to the total. Since PDB does not filter by room
status, we still are subject to a high number of rooms being evicted and system
highly impacted. To mitigate this, Maestro will now set a fixed amount for
maxUnavailable with default value to 5%, if not set on scheduler. It can be
configured via:

  MAESTRO_SERVICES_SCHEDULERMANAGER_DEFAULTPDBMAXUNAVAILABLE="10%"

The variable accepts integets greater than 0 and percentage strings.