fix(operations): ignore empty lease error when listing
When listing active operations, the internal handler fetches operation leases
to hydrate the operations with this info (useful for detecting when operation
is about to finish/expire). However, when listing the lease might have expired
thus it will throw an error that we don't want. Thus, ignoring lease not found
when listing active operations