fix(runtime-watcher): do not exit disruption on errors
Unerecoverable errors are normal when updating PDB, for instance
k8s might be updating at the same time, or we get throttled by
Kube API, or we simply fail to get from Redis. In either way, if
we return/exit we don't close the watcher since there's another
goroutine running, leaving to scenarios that watcher is running
only game room events loop. For now, if disruption failed, we log
and retry in the next loop