feat(policy): add fixed buffer autoscaling policy (#671)
* feat(policy): add new fixedBuffer autoscaling

The new policy will add a buffer of rooms of fixed size/amount
on top of the occupied rooms. Instead of relying on relative
occupancy to determine the amount of rooms in the system

* feat(api): new fixedBuffer policy to API

Managament API can receive a new policy type for scheduler autoscaling
which is the 'fixedBuffer' one. The validations were moved to a struct-level
validation so we can infere nested field values based on policy type

* chore(docker): remove no_cache from local dev
* fix(policy): refactor fixedBuffer map
* fix(metrics): default zero on game room metrics
* chore(worker): reduce log ingestion